scripts/validate_config.py
#!/usr/bin/env python3
"""
Network Configuration Validation Script
Validates network device configurations against defined standards
"""

import argparse
import json
import sys
from pathlib import Path

import yaml

def load_config_standards(standards_file):
    """Load configuration standards from YAML file"""
    try:
        with open(standards_file, 'r') as f:
            return yaml.safe_load(f)
    except Exception as e:
        print(f"Error loading standards file: {e}")
        return {}

def validate_cisco_config(config_data, standards):
    """Validate Cisco device configuration"""
    violations = []
    cisco_standards = standards.get('cisco', {})
    
    # Check for required global settings
    required_services = cisco_standards.get('required_services', [])
    for service in required_services:
        if service not in config_data:
            violations.append(f"Missing required service: {service}")
    
    # Check VLAN configurations
    vlan_standards = cisco_standards.get('vlans', {})
    if 'vlan_ranges' in vlan_standards:
        allowed_range = vlan_standards['vlan_ranges']
        # This would check actual VLAN configurations
        # Implementation would parse actual config data
    
    # Check interface naming conventions
    interface_standards = cisco_standards.get('interfaces', {})
    if 'description_required' in interface_standards:
        # Check that all interfaces have descriptions
        pass
    
    return violations

def validate_panos_config(config_data, standards):
    """Validate Palo Alto configuration"""
    violations = []
    panos_standards = standards.get('panos', {})
    
    # Check security zones
    required_zones = panos_standards.get('required_zones', [])
    for zone in required_zones:
        if zone not in config_data.get('zones', []):
            violations.append(f"Missing required security zone: {zone}")
    
    # Check logging configuration
    if panos_standards.get('logging_required', True):
        if not config_data.get('logging_enabled', False):
            violations.append("Logging is not enabled")
    
    return violations

def generate_compliance_report(validation_results):
    """Generate compliance report"""
    report = {
        'validation_timestamp': '{{ ansible_date_time.iso8601 }}',
        'total_devices': len(validation_results),
        'compliant_devices': 0,
        'non_compliant_devices': 0,
        'device_results': {}
    }
    
    for device, results in validation_results.items():
        is_compliant = len(results['violations']) == 0
        
        if is_compliant:
            report['compliant_devices'] += 1
        else:
            report['non_compliant_devices'] += 1
        
        report['device_results'][device] = {
            'compliant': is_compliant,
            'violations': results['violations'],
            'device_type': results.get('device_type', 'unknown')
        }
    
    return report

def main():
    """Main function"""
    parser = argparse.ArgumentParser(description='Network Configuration Validation')
    parser.add_argument('-s', '--standards', default='config_standards.yml',
                        help='Configuration standards file')
    parser.add_argument('-c', '--configs', required=True,
                        help='Directory containing device configurations')
    parser.add_argument('-o', '--output', default='compliance_report.json',
                        help='Output compliance report file')
    
    args = parser.parse_args()
    
    # Load configuration standards
    standards = load_config_standards(args.standards)
    if not standards:
        sys.exit(1)
    
    # Process configuration files
    config_dir = Path(args.configs)
    validation_results = {}
    
    for config_file in config_dir.glob('*.cfg'):
        device_name = config_file.stem
        
        # This is a simplified example - actual implementation would
        # parse device configurations using appropriate parsers
        config_data = {
            'device_type': 'cisco',  # Would be detected from config
            'services': [],  # Would be parsed from config
            'vlans': [],     # Would be parsed from config
            'zones': []      # For firewalls
        }
        
        # Validate based on device type
        if config_data['device_type'] == 'cisco':
            violations = validate_cisco_config(config_data, standards)
        elif config_data['device_type'] == 'panos':
            violations = validate_panos_config(config_data, standards)
        else:
            violations = ['Unknown device type - cannot validate']
        
        validation_results[device_name] = {
            'device_type': config_data['device_type'],
            'violations': violations
        }
    
    # Generate compliance report
    report = generate_compliance_report(validation_results)
    
    # Write report
    with open(args.output, 'w') as f:
        json.dump(report, f, indent=2)
    
    # Display summary
    print("Configuration Validation Complete")
    print("=" * 40)
    print(f"Total devices validated: {report['total_devices']}")
    print(f"Compliant devices: {report['compliant_devices']}")
    print(f"Non-compliant devices: {report['non_compliant_devices']}")
    print(f"Compliance rate: {(report['compliant_devices']/report['total_devices']*100):.1f}%")
    print(f"Detailed report saved to: {args.output}")

if __name__ == "__main__":
    main()
