requirements.txt
# Requirements file for network automation scripts
# Install with: pip install -r requirements.txt

# Core Ansible requirements
ansible>=4.0.0
ansible-core>=2.11.0

# Network device collections (install via ansible-galaxy)
# cisco.ios
# paloaltonetworks.panos  
# fortinet.fortios
# junipernetworks.junos
# arista.eos
# ansible.netcommon

# Python libraries for network automation
netaddr>=0.8.0          # IP address manipulation
textfsm>=1.1.0          # Parsing network device output
ntc-templates>=2.0.0    # TextFSM templates for network devices
pyyaml>=5.4.0           # YAML processing
jinja2>=2.11.0          # Template engine
requests>=2.25.0        # HTTP requests for API calls
urllib3>=1.26.0         # HTTP client library
cryptography>=3.4.0    # Cryptographic functions

# Device-specific libraries
paramiko>=2.7.0         # SSH client library
scp>=0.14.0             # SCP file transfer
pyserial>=3.5           # Serial port communication
netmiko>=3.4.0          # Multi-vendor network device library

# Palo Alto Networks
pan-os-python>=1.6.0    # PAN-OS API library
pandevice>=0.14.0       # PAN-OS device framework

# Juniper
junos-eznc>=2.6.0       # Juniper device automation library
jxmlease>=1.0.0         # XML to Python data structure conversion

# Fortinet  
fortiosapi>=1.0.0       # FortiOS API library

# Development and testing
pytest>=6.2.0           # Testing framework
pytest-cov>=2.12.0      # Coverage reporting
black>=21.0.0           # Code formatter
flake8>=3.9.0           # Linting
mypy>=0.910             # Type checking

# Documentation
sphinx>=4.0.0           # Documentation generator
sphinx-rtd-theme>=0.5.0 # ReadTheDocs theme

# Utility libraries
click>=8.0.0            # Command line interface creation
rich>=10.0.0            # Rich text and beautiful formatting
tabulate>=0.8.0         # Pretty-print tabular data
colorama>=0.4.0         # Cross-platform colored terminal text

# Optional: For advanced network analysis
scapy>=2.4.0            # Packet crafting and analysis
matplotlib>=3.4.0       # Plotting and visualization
pandas>=1.3.0           # Data analysis and manipulation

# Configuration management
python-dotenv>=0.19.0   # Environment variable management
configparser>=5.0.0     # Configuration file parsing
